scheme_env/server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
openenv-core
pydantic
